@@ -26,7 +26,8 @@
     "lint": "eslint \"**/*.{js,ts,tsx}\"",
     "prepare": "bob build",
     "release": "release-it",
-    "pods": "cd example/ios && pod-install --quiet"
+    "pods": "cd example/ios && pod-install --quiet",
+    "bootstrap": "yarn example && yarn && yarn pods"
   },
   "keywords": [
     "react-native",
@@ -59,14 +60,15 @@
     "eslint-plugin-ft-flow": "^2.0.3",
     "prettier": "^2.4.0",
     "react": "^18.2.0",
-    "react-native": "^0.71.6",
+    "react-native": "^0.72.7",
     "react-native-builder-bob": "^0.18.1",
     "release-it": "^14.11.5",
-    "typescript": "4.8.4"
+    "typescript": "^5.0.2"
   },
   "peerDependencies": {
     "react": "*",
-    "react-native": ">=0.71.0"
+    "react-native": "*",
+    "react-native-quick-base64": "*"
   },
   "release-it": {
     "git": {
@@ -165,7 +167,6 @@
     "@types/node": "^17.0.31",
     "crypto-browserify": "^3.12.0",
     "events": "^3.3.0",
-    "react-native-quick-base64": "^2.0.5",
     "stream-browserify": "^3.0.0",
     "string_decoder": "^1.3.0"
   }
