@@ -1,3 +1,4 @@
+const { getDefaultConfig, mergeConfig } = require('@react-native/metro-config');
 const path = require('path');
 const blacklist = require('metro-config/src/defaults/exclusionList');
 const escape = require('escape-string-regexp');
@@ -5,12 +6,9 @@ const pak = require('../package.json');
 
 const root = path.resolve(__dirname, '..');
 
-const modules = Object.keys({
-  ...pak.peerDependencies,
-});
+const modules = Object.keys({ ...pak.peerDependencies });
 
-module.exports = {
-  projectRoot: __dirname,
+const config = {
   watchFolders: [root],
 
   // We need to make sure that only one version is loaded for peerDependencies
@@ -23,18 +21,25 @@ module.exports = {
       )
     ),
 
-    extraNodeModules: modules.reduce((acc, name) => {
-      acc[name] = path.join(__dirname, 'node_modules', name);
-      return acc;
-    }, {}),
+    extraNodeModules: {
+      ...modules.reduce((acc, name) => {
+        acc[name] = path.join(__dirname, 'node_modules', name);
+        return acc;
+      }, {}),
+      // crypto: require.resolve('react-native-quick-crypto'),
+      // stream: require.resolve('stream-browserify'),
+      // buffer: require.resolve('@craftzdog/react-native-buffer'),
+    },
   },
 
   transformer: {
     getTransformOptions: async () => ({
       transform: {
-        experimentalImportSupport: true,
+        experimentalImportSupport: false,
         inlineRequires: true,
       },
     }),
   },
 };
+
+module.exports = mergeConfig(getDefaultConfig(__dirname), config);
